[Refactor] Split DriverState
