+++ be/src/exec/pipeline/pipeline_driver_queue.cpp
@@ -16,6 +16,7 @@
 
 #include "common/config_exec_flow_fwd.h"
 #include "compute_env/workgroup/work_group.h"
+#include "exec/pipeline/primitives/driver_state.h"
 #include "exec/pipeline/primitives/pipeline_metrics.h"
 #include "exec/pipeline/schedule/utils.h"
 #include "exec/pipeline/source_operator.h"
